requirements.txt
pandas==2.0.2
python-dotenv==1.0.0
python-dateutil==2.8.2
numpy==1.25.0
openpyxl==3.1.2
et-xmlfile==1.1.0